requirements.txt
asgiref==3.4.1
dj-rest-auth==2.2.2
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
